# contracts/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Robe.AI — Wardrobe Style API",
    "description": "Garment trait extraction, wardrobe management, style profiling, and recommendations.",
    "version": "v1"
  },
  "paths": {
...
    "/api/users/me/recommendations": {
      "post": {
        "tags": [
          "Recommendations"
        ],
        "summary": "Get ranked recommendations based on the user's style profile and inventory.",
        "description": "Requires a generated style profile. Returns items ranked by fit score,\r\nfiltered by optional budget, category, and occasion constraints.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecommendRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RecommendRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RecommendRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendationsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ColorInfo": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "hex": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "GarmentCategory": {
        "enum": [
          "top",
          "bottom",
          "dress",
          "outerwear",
          "footwear",
          "accessory",
          "other"
        ],
        "type": "string"
      },
      "GarmentFit": {
        "enum": [
          "slim",
          "regular",
          "loose",
          "oversized"
        ],
        "type": "string"
      },
...
      "GarmentPattern": {
        "enum": [
          "solid",
          "striped",
          "plaid",
          "checked",
          "floral",
          "graphic",
          "other"
        ],
        "type": "string"
      },
...
      "GarmentTraits": {
        "type": "object",
        "properties": {
          "category": {
            "$ref": "#/components/schemas/GarmentCategory"
          },
          "subcategory": {
            "type": "string",
            "nullable": true
          },
          "primaryColor": {
            "$ref": "#/components/schemas/ColorInfo"
          },
          "secondaryColors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ColorInfo"
            },
            "nullable": true
          },
          "pattern": {
            "$ref": "#/components/schemas/GarmentPattern"
          },
          "material": {
            "type": "string",
            "nullable": true
          },
          "fit": {
            "$ref": "#/components/schemas/GarmentFit"
          },
          "formality": {
            "type": "integer",
            "format": "int32"
          },
          "seasonality": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Season"
            },
            "nullable": true
          },
          "styleTags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "occasions": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "confidence": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
      "InventoryItem": {
        "type": "object",
        "properties": {
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "modifiedAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdByUserId": {
            "type": "string",
            "nullable": true
          },
          "modifiedByUserId": {
            "type": "string",
            "nullable": true
          },
          "id": {
            "type": "string",
            "nullable": true
          },
          "vendorId": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "traits": {
            "$ref": "#/components/schemas/GarmentTraits"
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "currency": {
            "type": "string",
            "nullable": true
          },
          "url": {
            "type": "string",
            "nullable": true
          },
          "imageUrl": {
            "type": "string",
            "nullable": true
          },
          "inStock": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
      "RecommendRequest": {
        "type": "object",
        "properties": {
          "maxBudget": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "currency": {
            "type": "string",
            "nullable": true
          },
          "categories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GarmentCategory"
            },
            "nullable": true
          },
          "occasion": {
            "type": "string",
            "nullable": true
          },
          "count": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "Recommendation": {
        "type": "object",
        "properties": {
          "inventoryItem": {
            "$ref": "#/components/schemas/InventoryItem"
          },
          "score": {
            "type": "number",
            "format": "double"
          },
          "reasoning": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "RecommendationsResponse": {
        "type": "object",
        "properties": {
          "recommendations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Recommendation"
            },
            "nullable": true
          },
          "profileGeneratedAt": {
            "type": "string",
            "format": "date-time"
          },
          "requestedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "Season": {
        "enum": [
          "spring",
          "summer",
          "fall",
          "winter",
          "all"
        ],
        "type": "string"
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Enter your Entra ID JWT token.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}